[feature] added actuator links, moved repositories, make fetchtype on user's role to eager, corrected getLong to parseLong in controller, better management of no bearer string, added ROLE_ACTUATOR role, added junit dependency , added some URL constants
