+++ pom.xml
@@ -113,6 +113,11 @@
             <artifactId>groovy</artifactId>
             <version>2.4.16</version>
         </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <scope>test</scope>
+        </dependency>
 
 
     </dependencies>
