+++ src/main/java/com/example/springsecurityauthtwo/security/model/entities/AppUser.java
@@ -30,7 +30,7 @@ public class AppUser {
     private String username;
     private String email;
     private String password;
-    @ManyToMany
+    @ManyToMany(fetch = FetchType.EAGER)
     @JoinTable(name = "user_roles",
             joinColumns = @JoinColumn(name = "user_id"),
             inverseJoinColumns = @JoinColumn(name = "role_id"))
