+++ src/main/java/com/example/springsecurityauthtwo/security/services/UserServicesImpl.java
@@ -6,8 +6,8 @@ import com.example.springsecurityauthtwo.security.model.dtos.SignupRequest;
 import com.example.springsecurityauthtwo.security.model.entities.AppRole;
 import com.example.springsecurityauthtwo.security.model.entities.AppUser;
 import com.example.springsecurityauthtwo.security.model.enumeration.ERole;
-import com.example.springsecurityauthtwo.security.model.entities.repositories.AppRoleRepository;
-import com.example.springsecurityauthtwo.security.model.entities.repositories.AppUserRepository;
+import com.example.springsecurityauthtwo.security.repositories.AppRoleRepository;
+import com.example.springsecurityauthtwo.security.repositories.AppUserRepository;
 import com.example.springsecurityauthtwo.security.tools.SecurityConstants;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
